Split file into two, because that what Java wants.
